Монах: Реализация способности "Буря, земля и огонь".
Монах: Переписана пассивная способность "Боевые рефлексы".
Монах: Переписан эффект оглушения способности "Неистовые кулаки".

+++ src/server/game/Entities/Unit/Unit.h
@@ -1363,7 +1363,7 @@ class Unit : public WorldObject
         void GetRandomContactPoint(const Unit* target, float &x, float &y, float &z, float distance2dMin, float distance2dMax) const;
         uint32 m_extraAttacks;
         bool m_canDualWield;
-        int32 countCrit;
+        float countCrit;
         uint8 insightCount;
 
         void _addAttacker(Unit* pAttacker)                  // must be called only from Unit::Attack(Unit*)
